requirements.txt
gspread
pandas
oauth2client
google-api-python-client
requests
beautifulsoup4
python-dotenv